+++ Changelog.md
@@ -18,3 +18,4 @@
 
 ## vitaminimporter 4.6
 * Retter SDM-13 - Primær nøgle skal hedde PID for registeret kan kopieres
+* Retter SDM-15 - Unikke nøgler på Indholdsstoffer og Udgaaede navne er ikke unikke